backend/requirements.txt
Flask
Flask-SQLAlchemy
Flask-Cors
openai
python-dotenv
